warnings: fix pedantic compiler warnings
Most of these are related to implicit sign conversion and width narrowing. They
have been "fixed" with casts in order to create a starting point for fixing
them correctly. Many of the changes in this commit are most definitely logically
_wrong_, but attempt to be consistent without breaking any functionality.

For example, the following old code:

  unsigned int a;
  int b = a;
  uint64_t c;
  uint32_t d = c;

Has become:

  unsigned int foo;
  int bar = (int)foo;
  uint64_t c;
  uint32_t d = (uint32_t)c;

This of course does nothing to fix any consequences of the narrowing and sign
conversions, but the casts become explicit.

By doing this as a first step, any attempt to fix the explicit casts more
correctly, while introducing a new warning, will be noticed by the pull-tester.