build: give warnings autoconf treatment as well
This allows for warnings to be added/removed even if they're only supported by
some compilers. Clang's -Weverything is a good example.

After turning on Clang's -Weverything, lots of new warnings are reported. They
have been disabled with the necessary flags, but it may be worth cleaning up
some of them and re-enabling those warnings.