secp256k1_ecdsa_sign: Document the low S value property.
For people using the library for Bitcoin, this is an important
guarantee to have on the signature, so it should be documented
and we should commit to it.

For people using the library for other purposes, the low S values
might be surprising since it makes this library behave differently
from the standard ECDSA signing algorithm.