Add speck logo
In preparation for the planned release (codename "Ivory Tower"), this
adds a carnivore-style logo. "secpk" is an anagram of "Speck", the
German word for bacon. The color of the fist is the orange used in
the Bitcoin logo.

The logo is based on two CC0-licensed images:
 - https://commons.wikimedia.org/wiki/File:CookedBacon.jpg
 - https://www.kisscc0.com/clipart/raised-fist-computer-icons-fist-bump-art-power-fis-b5f82z/

The font is Quantico.