Remove lax DER parsing functions
There is no need to have the same code in Bitcoin Core and its
subproject secp256k1, and this code duplication has led to divergent
revisions of the code.

This also moves the Travis script out of the contrib directory
because it was the only file left there and does not really inside
contrib anyway.