Introduce opaque secret key and stronger types
This changeset introduces a level of indirection between the EC secret key compact representation (a big endian 32 byte value) and the secp256k1_seckey type.  This allows for the type to have an internal representation that differs in the future.  Because the types are compatible today, this breaks the API but not the ABI.  Callers should parse their compact keys with secp256k1_ec_seckey_parse_compact before using that value where a secret key is expected.