Conditionally include stdio.h
The header `stdio.h` is only needed in code ifdef guarded by
`USE_EXTERNAL_DEFAULT_CALLBACKS` and also (indirectly) by `VERIFY`, we
can therefore guard the include statement in the same manner.

The reason for doing this as that wasm builds downstream have to patch
in an empty `stdio.h` file in order to build because of the
unconditional include.