tests: Make CHECK enforce 1 instead of just "!= 0"
This ensures that we don't omit the "== 1" in tests accidentally,
and thus also strengthens existing tests in which it has been omitted.

We want to check "== 1" in particular for the return values of API
functions, but it also makes sense in the case of internal functions.

If you really want to check "!= 0", you can still write it explicitly.