build: add back configure.ac change that was left out when integrating v0.3.0 ()
v0.3.0 slightly modified the way modules are enabled. We integrated v0.3.0 into
frost on 2023-11-21, with b701c4abaf0e, but we forgot to update the new
convention our FROST_SPECIFIC code in configure.ac. We do it now.

This is not a breaking change: both old and new versions should still be correct.