ci: enable FROST in the official CI pipeline
We enable Frost everywhere the official CI is enabling some optional modules. In
order to do this, we also have to set EXPERIMENTAL to true.

This change gives us the benefit of the extended coverage by the CI pipeline
introduced in 0.4.0 (ARM, s390, Windows, MacOS, MinGW). On the other side, the
duration for a complete CI run will increase significantly.

Closes #35.