frost: show that the frost header fails to be precompiled
This commit introduces a simple script that triggers the current failure to
precompile the frost header, due to a missing `uint32_t` symbol, and introduces
a CI workflow that verifies the failure (in order to fix it afterwards).

The script uses a convoluted syntax (taken from
https://gcc.gnu.org/bugzilla/show_bug.cgi?id=108732#c2) to trigger the header
precompilation without actually writing anything to disk. The reason is that for
header precompilation gcc 14.2 (2025-04-07), does not support `-o /dev/null`,
because it wants to mmap the output file.

In order to be as strict as possible, we compile in both C and C++, because the
headers for fixed length integers have different names in the two languages.

EVIDENCE OF THE ERROR (C):
    ```
    $ scripts/ensure-frost-header-is-precompilable.sh c
    INFO: Tryng to compile <BASE>/include/secp256k1_frost.h with c
    In file included from <command-line>:
    <BASE>/include/secp256k1_frost.h:19:5: error: unknown type name ‘uint32_t’
    19 |     uint32_t generator_index;
        |     ^~~~~~~~
    <BASE>/include/secp256k1_frost.h:12:1: note: ‘uint32_t’ is defined in header ‘<stdint.h>’; this is probably fixable by adding ‘#include <stdint.h>’
    11 | #include "secp256k1_extrakeys.h"
    +++ |+#include <stdint.h>
    12 |
    ```

EVIDENCE OF THE ERROR (C++):
    ```
    $ scripts/ensure-frost-header-is-precompilable.sh c++
    INFO: Tryng to compile <BASE>/include/secp256k1_frost.h with c++
    In file included from <command-line>:
    <BASE>/include/secp256k1_frost.h:19:5: error: ‘uint32_t’ does not name a type
    19 |     uint32_t generator_index;
        |     ^~~~~~~~
    <BASE>/include/secp256k1_frost.h:12:1: note: ‘uint32_t’ is defined in header ‘<cstdint>’; this is probably fixable by adding ‘#include <cstdint>’
    11 | #include "secp256k1_extrakeys.h"
    +++ |+#include <cstdint>
    12 |
    ```