cmake: Do not hard-code `CMAKE_CONFIGURATION_TYPES`
The initial idea was to place the desired default build configuration
first in `CMAKE_CONFIGURATION_TYPES`. The "Ninja Multi-Config" generator
treats the first configuration as the default, but this is not the case
for other generators.

This change uses the dedicated `CMAKE_DEFAULT_BUILD_TYPE` variable for
the same purpose, avoiding hard-coding `CMAKE_CONFIGURATION_TYPES` and
preserving users' settings, if any.