silentpayments: receiving
Add routine for scanning a transaction and returning the necessary
spending data for any found outputs. This function works with labels via
expliticly passing in a list of (label, label_tweak) entries and
needs access to the transaction outputs.
Requiring access to the transaction outputs is not suitable for light
clients, but light client support is provided in a future release.

Add an opaque data type for passing around the prevout public key sum
and the input hash tweak (input_hash). This data is passed to the scanner
before the ECDH step as two separate elements so that the scanner can
multiply the scan_key * input_hash before doing ECDH.

Finally, add test coverage for the receiving API.